centered icons, still no scrolls
