--- src/components/Mag/styles/mag.css
@@ -1,51 +1,0 @@
-.magazineCovers {
-  height: 300px;
-  width: auto;
-}
-
-.magazineCovers:hover {
-  -webkit-transform: scale(1.4, 1.4) rotate(20deg); /* Safari */
-}
-
-.panel {
-  background-color: transparent;
-}
-
-
-
-.bkgImage {
-  min-height: 100%;
-  min-width: 1024px;
-  width: 100%;
-  height: auto;
-  position: fixed;
-  top: 0;
-  left: 0;
-  z-index: -1;
-  -webkit-filter: blur(5px);
-  -webkit-transition: -webkit-filter .50s;
-}
-
-
-.row {
-   width: 80%;
-    margin-top:90px;
-    margin-left: 10%; 
-    margin-right: auto;
-    text-align: justify;
-    -ms-text-justify: distribute-all-lines;
-    text-justify: distribute-all-lines;
-}
-
-.row img {
-    vertical-align: top;
-    display: inline-block;
-    display: inline;
-  }
-
-  .stretch {
-    width: 100%;
-    display: inline-block;
-    font-size: 0;
-    line-height: 0
-}
