+++ src/components/MagPage/styles/magazinePage.css
@@ -1,0 +1,28 @@
+.bkgImage {
+  min-height: 100%;
+  min-width: 1024px;
+  width: 100%;
+  height: auto;
+  position: fixed;
+  top: 0;
+  left: 0;
+  z-index: -1;
+  -webkit-filter: blur(5px);
+  -webkit-transition: -webkit-filter .50s;
+}
+
+.magazineCovers {
+  height: 200px;
+  width: auto;
+  display : block;
+  margin : auto;
+}
+
+.magazineCovers:hover {
+  -webkit-transform: scale(1.4, 1.4) rotate(20deg); /* Safari */
+}
+
+.panel {
+  width: 90%;
+  margin-left: 5%;
+}
