+++ src/routes.js
@@ -2,13 +2,13 @@ import React from 'react'
 import { Route, IndexRoute } from 'react-router'
 import App from './components/App';
 import HomePage from './components/HomePage/HomePage';
-import MagazinesPage from './components/Mag/MagazinesPage';
+import MagPage from './components/MagPage/MagPage';
 import NotFoundPage from './components/NotFoundPage';
 
 const routes = (
 	<Route path="/" component={App}>
 		<IndexRoute component={HomePage} />
-		<Route path="magazines" component={MagazinesPage} />
+		<Route path="magazines" component={MagPage} />
 		<Route path="*" component={NotFoundPage} />
 	</Route>
 );
