+++ src/static/css/base.css
@@ -9,7 +9,8 @@
 }
 
 body {
-	height: 100%;
+	min-height: 100vh;
+	height: auto;
 	background-color: white;
     overflow: hidden;
     -webkit-background-size: cover;
